requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cssselect==1.2.0
DataRecorder==3.4.11
ddddocr==1.4.11
DownloadKit==2.0.0
DrissionPage==4.0.0b36
et-xmlfile==1.1.0
filelock==3.13.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.6
lxml==5.1.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
numpy==1.26.3
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.25.1
psutil==5.9.7
psycopg2==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.0
pynput==1.7.6
pyperclip==1.8.2
pyreadline3==3.4.1
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
schedule==1.2.1
six==1.16.0
sympy==1.12
tldextract==5.1.1
tzdata==2023.4
urllib3==2.1.0
websocket-client==1.7.0
